Fix access of private members in finally().
tag_invoke() methods are friends of the receiver but are
not transitively friends of the operation-state so
trying to access private members of the operation
class is not valid from a tag_invoke() method.

This was failing to compile under GCC, which is
stricter on friendship and access control than clang.