Simplify/optimise new_thread_context with use of atomics
The active-thread count is now an atomic variable.

No longer need to acquire the context lock to increment
the count when starting a new thread. Just do a relaxed
atomic increment instead.

Initialise the active-thread count to 1 and decrement
the count in the destructor so that retiring threads
don't call cv.notify_one() unless the destructor has
been entered.