Fix error mixing write_new_file() with sequence() in io_uring test.
The write_new_file() operation was producing the result of the
when_all() which would have been a pack of the individual
results of the async_write_some_at() calls.

This was causing problems when used within the sequence()
algorithm which now requires that non-last senders produce
an empty value pack.

We now explicitly discard the value results of the when_all()
oepration so that the write-operation becomes a void-sender.