Modify unifex::on() algorithm so it only stores scheduler once.
Previous implementation required both the with_query_value()
and the schedule() operation to hold a copy of the scheduler.

By putting the with_query_value() as the outer sender we can
then use the parametereless schedule() operation which uses
this context and avoid needing to store the scheduler twice.

This also allows us to then move the scheduler instead of copying.