Work around msvc eagerness to instantiate tag_invoke overloads in sequence().
Make CPO and sender/receiver arguments deduced and
then constrain to SFINAE out if the types don't match exactly.

Also modify handling of sequence() overload for 3 or more
parameters to avoid infinite recursion of function instantiation
under msvc by adding an explicit third argument so that the
overload would not be considered for the recrusive call to
sequence() with two argments.