Make transform/when_all connect() methods SFINAE-friendly.
Calling connect() on a transform/when_all sender was attempting to
instantiate both rvalue-ref and const-lvalue-ref overloads of connect().
If the predecessor was only rvalue-connectable then
this would result in a compile-error.

This change modifies the connect() methods to SFINAE out if that particular connect() overload is not supported.

This fixes compilation errors in io_uring_test that were
introduced in #95 which added some const-lvalue-ref
overloads of connect().