Reduce instantiations of connect() overloads.
Implement the connect() methods for the let, sequence, transform and
when_all algorithms by defining them using tag_invoke and templating
them on the sender type so that we can be deterministic about which
overload is instantiated.

When we had two connect() method overloads (const-ref and rvalue-ref)
the compiler would often try to instantiate both of them during overload
resolution, even if the sender was an rvalue. When these were composed
this would lead to a combinatorial explosion of connect() method
instantiations during overload resolution of O(2^depth) instantiations.

This change reduces compile-time of the io_uring_test.cpp file by
about 2/3.