Improve compile-time of sync_wait().
The sync_wait() implementation had an optimisation for senders that complete
synchronously that would avoid the use of synchronisation. However, it did
this by instantiating the sender's operation with two different receivers;
one that incorporated synchronisation and one that did not. This meant that
the compiler had to instantiate the entire operation chain multiple times
which could be quite expensive if the sender was a complex composed
operation.

We now just instantiate the operation once and use a branch in the receiver
to indicate whether to perform the synchronisation or not. This should
halve the compilation cost of sync_wait().