Remove support for extra Args... forwarding on receiver queries.
This reduces the number of template instantiations by avoiding
instantiating these queries when resolving connect() overloads.

This ability to pass extra args is currently unused but if we need
to add it back in later we would need to constrain the CPO to
filter out the 'connect' CPO to avoid these overloads being
considered for 'connect'.