Add unifex/win32/detail/types.hpp to avoid needing to include Windows.h publically
This header defines some common Win32 types and structures that we
want to use within various windows executor implementations.

Defining the types here allows us to use them in the public interface
of the executor types and their sender operations without needing
to #include <Windows.h> in a public header.