Add initial version of unifex::win32::low_latency_iocp_context.
This is a proof-of-concept design of a scheduler that tries to avoid
always needing two syscalls to perform an async I/O operation for
operations that complete quickly with low-latency by introducing
an extra polling step that polls for I/O completion of recently started
I/O operations before going back to GetQueuedCompletionStatus().

If the operation completes as part of the polling then we call the
receiver and detach from the OVERLAPPED strructures which will
be garbage collected later when the completion-events are received
in a subsequent call to GetQueuedCompletionStatus().