Introduce unifex::type_index to support -fno-rtti
This change introduces <unifex/type_index.hpp> and unifex::type_index to
support building Unifex without RTTI.

When RTTI is enabled unifex::type_index is an alias for std::type_index.
When RTTI is disabled unifex::type_index is a struct with a work-alike
interface to std::type_index. In either mode, unifex::type_id<T>() is a
substitute for typeid(T).