Suppress vectorization warning
When building with Clang with -Oz, the compiler will not vectorize the
loops in bulk_schedule.hpp that have requested vectorization.  This
results in a warning like so:

```
In file included from libunifex/test/bulk_schedule_test.cpp:17:
libunifex/include/unifex/bulk_schedule.hpp:96:17: error: loop not
vectorized: the optimizer was unable to perform the requested
transformation; the transformation might be disabled or specified as
part of an unsupported transformation ordering
[-Werror,-Wpass-failed=transform-warning]
                for (Integral i(0); i < count_; ++i) {
                ^
1 error generated.
```

This change suppresses the warning with a Clang-specific #pragma.