Start to support -fno-exceptions (#160)
* Convert retry_when_test.cpp to use GTest

The test in test/retry_when_test.cpp just exercised retry_when in a
custom int main() and reported results with return codes.  This change
converts the test to use Google Test, like the rest of the tests.

* Replace function-try-blocks

Change this:

    void foo() noexcept try {
      // ...
    } catch(...) {
      // ...
    }

to the equivalent

    void foo() noexcept {
      try {
        // ...
      } catch(...) {
        // ...
      }
    }

It's going to make it easier to conditionally compile out the try and
catch keywords without duplicating code.

* Mostly support compiling without exceptions

This change introduces UNIFEX_NO_EXCEPTIONS and then uses it to
conditionally compile out the try keyword and catch blocks.  There are
still some throw statements in the source where Unifex throws exceptions
to report errors; we'll need to conditionally compile those out, too, to
fully support compiling without exceptions, but that will require some
discussion about what to do instead.

* Add UNIFEX_TRY, UNIFEX_CATCH, and UNIFEX_RETHROW

Take @ericniebler's advice and switch to macros that replace `try` and
`catch`.  I had to come up with a replacement for `throw;` as well,
because `throw;` in a not-taken branch of an `if constexpr` still trips
up the compiler when compiling with `-fno-exceptions`.