Add inplace_stop_token_adapter.
This is a helper that allows adapting any incoming StopToken
type to forward its signal to an inplace_stop_token for passing
through type-erased wrappers that require this type.

Includes some specialisations that optimise for unstoppable_token
and for the case where the input StopToken type is already an
inplace_stop_token.