Rework cancellation support for coroutines.
Coroutines now unconditionall unwind themselves if an awaited
operation completes with 'done'.

This signal will be transported up the stack until it hits a consumer
that does handle the 'done' signal. This typically means wrapping
the task<T> in a sender-algorithm that translates the done-signal
into some other signal that a coroutine can then handle.
eg. as an error/exception or as a value.

This means we don't have to unwrap std::optional values every
time we co_await something.

This relies on coroutine promise_types implementing a new
'unhandled_done()' method, similar to 'unhandled_exception()',
that handles the 'done' signal for that particular coroutine type.

The 'task' type simply forwards the signal on to the awaiting
coroutine's promise.unhandled_done() method. This recursion
is terminated by the connect_awaitable() implementation which
implements unhandled_done() to forward on to the receiver's
set_done() implementation.