Introduce a new synchronization primitive
I don't know what to name this thing so it's called `unnamed_primitive`
for now.  I think it might be some kind of semaphore, but I'm not sure.

The semantics are very similar to `async_manual_reset_event` but:
 - there may only be one waiter at a time; trying to add a second will
   terminate the process
 - waiters may cancel their wait

The method names are the same as `async_manual_reset_event`:
 - async_wait() to get a _Sender_ upon which to wait
 - set() to signal waiter to wake up
 - reset() to return to the unsignalled state
 - ready() to query whether the object is currently signalled

I've chosen to `std::terminate()` when I detect an incorrect state; I
don't know whether this is acceptable.  We could throw an exception,
instead, I suppose, but the sorts of errors I'm detecting are
unrecoverable programmer errors so I've erred on the side of failing
fast.