Introduce unifex::async_scope (#237)
`async_scope` is a type for collecting async work you'd like to almost
fire-and-forget, but not actually forget.  It has a `spawn` method that
takes a _Sender_ and a _Scheduler_, connects the _Sender_ to an internal
_Receiver_ that drops results on the floor, and arranges for the
resulting operation state to be stored on the heap so it can be started
without having to keep the starting function's stack frame alive.