get_stop_token: add compile-guard to deduction guide
GCC fails to compile the deduction guide with:
```
    error: deduction guide 'unifex::_get_stop_token::_fn::_awaitable::_awaiter(StopToken) -> unifex::_get_stop_token::_fn::_awaitable::_awaiter' must be declared at namespace scope
```

This has already been reported upstream with gcc issue 79501 in 2017 and
is still unresolved.  The deduction guide seems unnecessary in light of
P0091 and everything compiles on gcc-10 without it.

clang-12 on macOS requires it or else it fails with:
```
    ../source/../include/unifex/get_stop_token.hpp:50:16:
    error: no viable constructor or deduction guide for deduction of template arguments of '_awaiter'
    ...
    ../test/task_cancel_test.cpp:77:12:
    error: no member named 'await_ready' in 'unifex::_get_stop_token::_fn::_awaitable'
    auto a = co_await get_stop_token();
```

Add a compile guard so the deduction guide is present on clang and not
present on GCC.

Fixes facebookexperimental/libunifex#284.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>