config: rework UNIFEX_NO_UNIQUE_ADDRESS for gcc-10
GCC 10.3 introduced bug 98995 (also in 11+), which breaks
compilation when using [[no_unique_address]] due to missing
copy-elision.  The only viable solution for now is to re-define
UNIFEX_NO_UNIQUE_ADDRESS as empty.

There was already a special case for certain versions of clang in
the UNIFEX_NO_UNIQUE_ADDRESS #define.  Rewrite it for clarity
and inclusion of the workaround for the gcc-10+ series.

Even though 10.[0-2] do not have bug 98995, the users of those
versions and the scope of optimization introduced by using
the [[no_unique_address]] attribute are likely minimal enough that
I opted for simplicity of the #define tree and determined any GCC-10
or above currently needs the workaround.

Resolves facebookexperimental/libunifex#307.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>