Adapt unifex::create and its tests to broken-stdlib
The unit tests for `unifex::create` make direct mention of
`std::optional`, which ought to be redirected through `unifex::optional`
in the `broken-stdlib` branch.

`unifex::create` also uses the `_v`-form of some type traits that need
to be redirected from `std::`-qualified to the `unifex::` version.

This diff does both sets of redirections.