stop_immediately: avoid start_cleanup after destruction
When adding a virtual destructor to stop_immediately
cleanup_operation_base, test cases started to fail in GCC with runtime
errors calling a missing pure virtual function.  This was caused by the
destructor of the cleanup_sender being called twice: once by take_until
and once by its own start_cleanup.

Virtual classes are not allowed to have their destructor called twice.
The implementation in GCC seems to set the vtable to the inherited
classes as the destructors are executed, such that when the last
destructor is called the vtable points to the base class with only
pure virtual functions.  Any further operation on the class results
in missing virtual functions.

This code is likely terribly wrong but cleans up the test case failures
and successfully obtains the value/error/done state from the
stop_immediately.  I would love to have pointers on how to deal with
this properly.  There is already a 'cleanup_completed' state which is
completely unused, so it seems like there was an oversight in the
implementation somewhere.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>