Fix ownership in async_scope::detached_spawn (#376)
PR #372 has a subtle ownership bug that's mostly attributable to a lack
of tests for `detached_spawn` and friends.  If `async_scope::cleanup()`
triggers `set_done()` on a detached promise because of the stop callback
registered on the scope's stop source, we'll delete the operation state
while using its `inplace_stop_source`, leading to a use-after-free.

This diff fixes the problem by making the guilty stop callback a
part-owner in the operation state.  There are now three owners: the
`future<>` (if there is one), the operation itself, and the stop
callback that propagates stop requests from the scope's stop source to
each operation's stop source.

I'll figure out how to write a unit test that captures the now-fixed
failure mode in a future PR.