Minimal changes to allow compilation to succeed on gcc and clang on MSYS2 (#391)
* Typo fix (_Out_writes_to -> _Out_writes_to_)

* Include <cstring> for std::memcpy

* Avoid redefining _cdecl and __cdecl on MinGW

* Mark nameless union as an extension on g++

This avoids a warning under -pedantic.

* Avoid 'unused' errors on g++

Mark variables/bindings as "maybe unused" and delete parameter names.

* Cast function pointer from FARPROC via void(*)()

This suppresses the -Wcast-function-type warning, as mentioned in the
documentation for that option.

* Add missing override directives to suppress warnings

* Add nested class aliases to private types in enclosing class

On g++ a function that is a friend of a nested class cannot access private
member types of the enclosing class (!)

* windows_thread_pool::schedule_sender::connect constraints: fix typo

"is_stop_never_possible_v<Receiver>"
should be
"is_stop_never_possible_v<stop_token_type_t<Receiver>>".

* CI: Pin MSVC version to 2019

Co-authored-by: Richard Copley <buster@buster.me.uk>