Allow any_object to contain types with throwing move constructors.
If any_object has 'RequireNoexceptMove' as 'true' and the type being
stored in the any_object does not have a nothrow move-constructor then
it falls back to heap-allocating the type instead of storing it in the
inline storage so that the any_object move constructor can still be
marked noexcept (moving a heap-allocated object never throws since it
is just moving ownership of the pointer).

Also added support for an assignment operator. This required adding
support for an invalid state to handle the case where we fail to
construct the new value after having already destroyed the existing
value.

The 'with_abort_tag_invoke' helper assists with defining an 'invalid_obj'
type that customises all CPOs (and thus allows generating a valid vtable)
but where every CPO (other than destructor and move-ctor) end up calling
std::abort(). The invalid state is represented by effectively initialising
the any_object to hold an instance of this type.