Fix issue with any_object under MSVC which was causing cl.exe memory exhaustion
The any_object(std::allocator_arg, alloc, std::in_place_type<T>, args...) constructor
was delegating to a call to another any_object constructor to construct an inline
instance of type detail::any_heap_allocatoed_storage.

Unfortuantely, MSVC seems to get itself into an infinite recursion of calls to
constructors of different template instantiations and does so until it runs out
of memory.

The workaround applied here is to manually duplicate the body of the
constructor we intended to call instead of calling the other constructor.