Check receiver earlier to avoid gcc11+ compiler error
gcc11 and above seems to be more eager in instantiating templates in
evaluating `is_tag_invocable`, and can fail compilation rather than
evaluating `is_tag_invocable` as false.

Test plan: The new test case fails to compile on gcc 11, 12, but passes
with gcc 10 and clang 13 and 14. The fix to let_error.hpp fixes the test
to compile and pass.

Fixes #422