Disable coroutine support for GCC 10.2 and earlier
GCC 10.2 coroutine support is not great. In particular,
this version of GCC does not handle promise type constructors
that accept the coroutine function parameters correctly. GCC
passes a reference to the pre-copy object to the promise type
constructor rather than a reference to the coroutine function
argument after it's been copied to the coroutine frame.
connect_awaitable relies on this working correctly, so we will
require GCC 10.3 or newer to use coroutines with unifex.