Work around an MSVC bug in C++20 mode (#492)
* Work around an MSVC bug in C++20 mode

PR #489 seems to have found a bug in MSVC when compiling
`variant_sender` in C++20 mode.  This diff adds a test that I expect to
pass everywhere but fail to build with MSVC.  The next diff will fix the
problem with a work-around.

* Actually work around MSVC's bug

This diff works around the bug revealed by the test added in the
previous diff.