Simplify v2::async_scope's nest_sender::operator=
This diff simplifies the implementation of `operator=` for
`v2::async_scope`'s `nest_sender`.  The new implementation is less
branchy, and easier to understand (I hope).  I'm a little uncomfortable
with the fact that `operator=` is now always implemented in terms of the
nested Sender's move-constructor instead of move-assignment operator,
but the trade-off seems worth it.