Fix regression in task<>
Prior to #290, task<> could co_await awaitables that returned references
in await_resume.

Test plan: added test case that passes at the commit before #290, then
fails after #290, and passes again with the changes in this commit.
I also added a test for task<> awaiting a sender that sends a
reference, but this was not actually broken.