Update finally to support value channel references
This aligns with the general idea that senders should be able to send
references and values, so the adaptor algorithms should preserve the
types in value channels when there is no specific requirement to modify
the value channel type.

In making this change, the CreateTest.AwaitTest test failed to compile.
This appears to be the create operation accepting universal reference
`Ts&&...` arguments which are then forwarded to the underlying receiver's
`set_value`. However, in this test, the type of `result` is `int&`,
but the create sender only sends `int`. Before this change (and before
the schedule affine task), something else ended up decaying the
parameter's type further downstream. With the scheduler affine task
and its reliance on `finally`, the `finally` receiver's `set_value`
strictly enforces that it's only called with the expected type, which
caused the compilation error with my changes. Fundamentally, the create
sender was incorrectly sending `int&` despite advertising it only
sends `int`. I updated the create algorithm to faithfully send exactly
what it advertises to fix the problem.

I left finally's existing behavior of decay_t-ing error types since I
ran into other issues trying to get that to work, and it seems less
useful/likely (though, I don't see why it shouldn't be done later).

Note that per discussion in #541, this is a source and behavior
breaking change.