Rename typed_bulk_sender to bulk_sender
In the spirit of PR #539 and the previous diff, this diff banishes the
concept of typed bulk senders by renaming `unifex::typed_bulk_sender<>`
to `unifex::bulk_sender<>`. There's no such thing as an untyped bulk
sender so there's no need to point out that a bulk sender is typed.