Create a way to disable let_value's cleanup asserts
Introduce `UNIFEX_DISABLE_LET_VALUE_CLEANUP_ASSERTS` as a build-time
switch to control whether `let_value`'s operation state makes assertions
about the identity of its cleanup function. This addresses an issue
where instantiating a `let_value` operation state in one shared library
and completing it in another results in false-positive asserts because
the cleanup function that will be invoked is the correct one, but in a
different shared library than the one evaluating the assert.