Add support for function references (free and static member) in bind_back (#605)
* Use std::is_invocable variants

These are drop in replacements of is_callable variants and we can
replace our concept shim with them as well.

* Use ASSERT_TRUE with optional/expected type has_value assertions

These assertions are usually followed by other expressions
assuming that they are passing and in case a change breaks them
they potentially trigger undefined behavior with `EXCEPT_TRUE`.

* Add support for function references (free and static member) in bind_back

The following program does not compile: https://godbolt.org/z/WfhEc51dY
But if we replace the function reference with a lambda, it does: https://godbolt.org/z/jrhTY4nse
Constructing a std::tuple with a function type fails but function pointers are allowed.
We instantiate _result with std::decay_t<ArgN>... which maps function references to function pointers
and then the corresponding std::tuple instantiation does not fail to compile. The problem with the
existing code is that, in order to instantiate std::tuple<std::decay_t<ArgN>...>, we first constructed a
std::tuple<ArgN> and passed it over. If instead we construct a std::tuple<std::decay_t<ArgN>...>
with the function arguments, they will be converted to function pointers and the std::tuple instantiation
will compile.