Add ScopedAsyncStackRoot::ensureFrameDeactivated()
We need a way to restore a `ScopedAsyncStackRoot` to the "no active
frame" state before destroying it on the way out of a customization of
`unifex::start` but the frame we want to deactivate is a member of the
operation state, which means it's likely already been destroyed. This
diff adds `ScopedAsyncStackRoot::ensureFrameDeactivated()`, which
performs most of the same actions as `deactivateAsyncStackFrame()` but
without touching the frame. I think this still technically invokes UB by
copying and comparing a zapped pointer, but it's better than what we had
before.