Fix CPO forwarding in upon_done and _error
The existing CPO-forwarding implementations of `tag_invoke` in
`unifex::upon_done` and `unifex::upon_error` get the order of operations
incorrect when trying to cast the provided CPO to an rvalue, resulting
in trying to cast the *result* of the CPO to the CPO's type. This diff
fixes it by switching to `std::move(cpo)`, which automatically gets the
operator precedence correct.